refactor(ertp-ledgerguise): consolidate mock IO utilities
Create test/mock-io.ts with:
- makeTestClock: deterministic clock advancing by fixed step
- mockMakeGuid: sequential hex GUID generator
- makeTestDb: in-memory GnuCash database setup

Remove test/helpers/clock.ts and update imports.
